Correctly handle `order: :none` in semantic search
Setting `order: :none` in semantic search raised an error since when we've expanded
`EctoQuery.order_by/3` macro it checked for the validity of the `order` option.

Now this should be fixed and `order: :none` correctly omits ordering.

This should not affect any existing logic, since there is a little value to use
semantic search without ordering.